# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.14"
  title: DTL Service
# during dev, should point to your local machine
host: localhost:80
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - application/octet-stream #smell, this is necessary because supertest appears to send this content type for empty posts
# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/plain
  - text/html
paths:
  /findWpExperience/tree/{experienceId}:
    x-swagger-router-controller: WpController
    get:
      description: Get Experience Structure by Experience ID
      operationId: getExperienceStructure
      parameters:
      - name: experienceId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Experience
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EntityResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
  ### refers to MYSQL view that contains JOIN from experience to venue using 
  ### WP taxanomy structure that provides 
  ### a DIRECT experience to venues relationships to support geo-location 
  ### searches by geo location (lat lng) AND by venue_id
...
  /migrations/up:
    x-swagger-router-controller: MigrationsController
    get:
      description: Run all pending migrations
      operationId: migrateUp
      parameters:
        - name: token
          in: query
          type: string
          required: false
          description: Access token to run migrations
      responses:
        "200":
          description: Success
        "default":
          description: Error
          schema:
              $ref: "#/definitions/ErrorResponse"
  /migrations/down:
    x-swagger-router-controller: MigrationsController
    get:
      description: Run all pending migrations
      operationId: migrateDown
      parameters:
        - name: token
          in: query
          type: string
          required: false
          description: Access token to run migrations
      responses:
        "200":
          description: Success
        "default":
          description: Error
          schema:
              $ref: "#/definitions/ErrorResponse"
...
  /seriesalliances/series/{seriesId}:
    x-swagger-router-controller: WpController
    get:
      description: Get Alliances by Series ID
      operationId: getSeriesAlliance
      parameters:
      - name: seriesId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Series
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AllianceListResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
  /eventsbyseries/series/{seriesId}:
    x-swagger-router-controller: WpController
    get:
      description: Get Events by Series ID
      operationId: getEventBySeriesId
      parameters:
        - name: seriesId
          type: number
          in: path
          required: true
          description: The unqiue numerical identifier for Series
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EventListResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"            
  /findAllAlliances/series/{seriesId}:
    x-swagger-router-controller: WpController
    get:
      description: Find Alliances by Series ID
      operationId: findAllAlliancesBySeries
      parameters:
      - name: seriesId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Series
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AllianceListResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
  /findAllAlliancesByExperience/experience/{experienceId}/series/{seriesId}:
    x-swagger-router-controller: WpController
    get:
      description: Find Alliances by Experience and Series ID
      operationId: findAllAlliancesByExperience
      parameters:
      - name: experienceId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Experience
      - name: seriesId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Series
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AllianceListResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
  /findAllAlliancesByExperience/experience/{experienceId}:
    x-swagger-router-controller: WpController
    get:
      description: Find Alliances by Experience ID
      operationId: findAllAlliancesByExperience
      parameters:
      - name: experienceId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Experience
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AllianceListResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
  /findAllSeries/experience/{experienceId}:
    x-swagger-router-controller: WpController
    get:
      description: Find Alliance by Series and User ID
      operationId: findAllSeriesByExperience
      parameters:
      - name: experienceId
        type: number
        in: path
        required: true
        description: get all associated series for experience
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Series"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
  /findExperience/slug/{slugId}:
    x-swagger-router-controller: WpController
    get:
      description: Find Experience by WP Experience Slug
      operationId: findExperienceBySlug
      parameters:
      - name: slugId
        type: string
        in: path
        required: true
        description: WordPress Experience Slug
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Experience"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"            
...
definitions:
...
  EntityResponse:
    type: object
  # Response definitions
  ErrorResponse:
    required:
      - type
      - code
      - message
    properties:
      type:
        type: string
      code:
        type: number
      message:
        type: string
  AllianceListResponse:
    type: array
    items: 
      $ref: "#/definitions/Alliance"
...
  EventListResponse:
    type: array
    items: 
      $ref: "#/definitions/Event"
...
  Alliance:
    type: object
...
  Event:
    type: object
  Experience:
    type: object
...
  Series:
    type: object
...